some fixes and changes and addition

+++ wawl/Console.hpp
@@ -36,7 +36,7 @@ namespace wawl {
 				bool isSuccess = ::WriteConsole(
 					consoleOutput,
 					str.c_str(),
-					str.size() + 1,
+					static_cast<Uint32>(str.size() + 1),
 					&writtenSize,
 					nullptr
 					) != 0;
